Move decode_*() functions from dwarf.rs to tools.rs.
These code are used by several modules.  They should be moved out of
dwarf.rs.
